+++ src/vikmapslayer_compat.h
@@ -29,7 +29,7 @@
 G_BEGIN_DECLS
 
 typedef struct {
-  guint8 uniq_id;
+  guint16 uniq_id;
   guint16 tilesize_x;
   guint16 tilesize_y;
   guint drawmode;
