+++ src/vikmapsourcedefault.c
@@ -35,7 +35,7 @@ static const gchar *map_source_get_license (VikMapSource *self);
 static const gchar *map_source_get_license_url (VikMapSource *self);
 static const GdkPixbuf *map_source_get_logo (VikMapSource *self);
 
-static guint8 map_source_get_uniq_id (VikMapSource *self);
+static guint16 map_source_get_uniq_id (VikMapSource *self);
 static const gchar *map_source_get_label (VikMapSource *self);
 static guint16 map_source_get_tilesize_x (VikMapSource *self);
 static guint16 map_source_get_tilesize_y (VikMapSource *self);
@@ -54,7 +54,7 @@ struct _VikMapSourceDefaultPrivate
 	gchar *license_url;
 	GdkPixbuf *logo;
 
-	guint8 uniq_id;
+	guint16 uniq_id;
 	gchar *label;
 	guint16 tilesize_x;
 	guint16 tilesize_y;
@@ -359,10 +359,10 @@ map_source_get_logo (VikMapSource *self)
 	return priv->logo;
 }
 
-static guint8
+static guint16
 map_source_get_uniq_id (VikMapSource *self)
 {
-	g_return_val_if_fail (VIK_IS_MAP_SOURCE_DEFAULT(self), (guint8)0);
+	g_return_val_if_fail (VIK_IS_MAP_SOURCE_DEFAULT(self), (guint16)0);
 	
 	VikMapSourceDefaultPrivate *priv = VIK_MAP_SOURCE_DEFAULT_PRIVATE(self);
