+++ src/vikslippymapsource.h
@@ -50,7 +50,7 @@ struct _VikSlippyMapSource
 
 GType vik_slippy_map_source_get_type (void) G_GNUC_CONST;
 
-VikSlippyMapSource * vik_slippy_map_source_new_with_id (guint8 id, const gchar *label, const gchar *hostname, const gchar *url);
+VikSlippyMapSource * vik_slippy_map_source_new_with_id (guint16 id, const gchar *label, const gchar *hostname, const gchar *url);
 
 G_END_DECLS
